backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pinecone
langchain-huggingface==0.0.1
pydantic
langchain

